Migrate fail_dir functions to module
