+++ makeflow/src/makeflow_gc.c
@@ -349,84 +349,4 @@ int makeflow_clean_mount_target(const char *target) {
 	return 0;
 }
 
-int makeflow_clean_rm_fail_dir(struct dag *d, struct dag_node *n, struct batch_queue *q) {
-	assert(d);
-	assert(n);
-	assert(q);
-
-	int rc = 0;
-	char *faildir = string_format(FAIL_DIR, n->nodeid);
-	struct dag_file *f = dag_file_lookup_fail(d, q, faildir);
-	if (!f) goto OUT;
-
-	if (makeflow_clean_file(d, q, f, 1)) {
-		debug(D_MAKEFLOW_RUN, "Unable to clean failed output");
-		goto OUT;
-	}
-
-	rc = 1;
-
-OUT:
-	free(faildir);
-	return rc;
-}
-
-int makeflow_clean_prep_fail_dir(struct dag *d, struct dag_node *n, struct batch_queue *q) {
-	assert(d);
-	assert(n);
-	assert(q);
-
-	int rc = 1;
-	char *faildir = string_format(FAIL_DIR, n->nodeid);
-	struct dag_file *f = dag_file_lookup_fail(d, q, faildir);
-	if (!f) goto FAILURE;
-
-	if (makeflow_clean_file(d, q, f, 1)) {
-		debug(D_MAKEFLOW_RUN, "Unable to clean failed output");
-		goto FAILURE;
-	}
-	if (batch_fs_mkdir(q, f->filename, 0755, 0)) {
-		debug(D_MAKEFLOW_RUN, "Unable to create failed output directory: %s", strerror(errno));
-		goto FAILURE;
-	}
-
-	makeflow_log_file_state_change(d, f, DAG_FILE_STATE_COMPLETE);
-	fprintf(stderr, "rule %d failed, moving any outputs to %s\n",
-			n->nodeid, faildir);
-	rc = 0;
-FAILURE:
-	free(faildir);
-	return rc;
-}
-
-int makeflow_clean_failed_file(struct dag *d, struct dag_node *n,
-		struct batch_queue *q, struct dag_file *f, int prep_failed,
-		int silent) {
-	assert(d);
-	assert(n);
-	assert(q);
-	assert(f);
-
-	if (prep_failed) goto CLEANUP;
-
-	char *failout = string_format(
-			FAIL_DIR "/%s", n->nodeid, f->filename);
-	struct dag_file *o = dag_file_lookup_fail(d, q, failout);
-	if (o) {
-		if (batch_fs_rename(q, f->filename, o->filename) < 0) {
-			debug(D_MAKEFLOW_RUN, "Failed to rename %s -> %s: %s",
-					f->filename, o->filename, strerror(errno));
-		} else {
-			makeflow_log_file_state_change(d, f, DAG_FILE_STATE_DELETE);
-			debug(D_MAKEFLOW_RUN, "Renamed %s -> %s",
-					f->filename, o->filename);
-		}
-	} else {
-		fprintf(stderr, "Skipping rename %s -> %s", f->filename, failout);
-	}
-	free(failout);
-CLEANUP:
-	return makeflow_clean_file(d, q, f, silent);
-}
-
 /* vim: set noexpandtab tabstop=4: */
