+++ makeflow/src/makeflow_module_fail_dir.c
@@ -1,0 +1,211 @@
+
+#include "makeflow_hook.h"
+#include "xxmalloc.h"
+#include "debug.h"
+#include "stringtools.h"
+#include "makeflow_gc.h"
+#include "makeflow_log.h"
+#include "dag.h"
+#include "dag_node.h"
+#include "dag_file.h"
+#include "jx.h"
+
+#include <assert.h>
+#include <unistd.h>
+#include <errno.h>
+#include <signal.h>
+#include <stdio.h>
+#include <stdlib.h>
+#include <string.h>
+#include <stdbool.h>
+
+#define FAIL_DIR "makeflow.failed.%d"
+
+struct dag_file *dag_file_lookup_fail(struct dag *d, struct batch_queue *q, const char *path) {
+	assert(d);
+	assert(q);
+	assert(path);
+	struct stat buf;
+	struct dag_file *f = dag_file_from_name(d, path);
+	if (f) {
+		if (f->type == DAG_FILE_TYPE_INPUT) {
+			debug(D_MAKEFLOW_HOOK,
+					"skipping %s since it's specified as an input",
+					path);
+			return NULL;
+		}
+		return f;
+	} else {
+		if (!batch_fs_stat(q, path, &buf)) {
+			debug(D_MAKEFLOW_RUN,
+					"skipping %s since it already exists",
+					path);
+			return NULL;
+		}
+		return dag_file_lookup_or_create(d, path);
+	}
+}
+
+int makeflow_clean_rm_fail_dir(struct dag *d, struct dag_node *n, struct batch_queue *q, struct makeflow_alloc *alloc) {
+	assert(d);
+	assert(n);
+	assert(q);
+
+	int rc = 0;
+	char *faildir = string_format(FAIL_DIR, n->nodeid);
+	struct dag_file *f = dag_file_lookup_fail(d, q, faildir);
+	if (!f) goto OUT;
+
+	if (makeflow_clean_file(d, q, f, 1, alloc)) {
+		debug(D_MAKEFLOW_RUN, "Unable to clean failed output");
+		goto OUT;
+	}
+
+	rc = 1;
+
+OUT:
+	free(faildir);
+	return rc;
+}
+
+int makeflow_clean_prep_fail_dir(struct dag *d, struct dag_node *n, struct batch_queue *q, struct makeflow_alloc *alloc) {
+	assert(d);
+	assert(n);
+	assert(q);
+
+	int rc = 1;
+	char *faildir = string_format(FAIL_DIR, n->nodeid);
+	struct dag_file *f = dag_file_lookup_fail(d, q, faildir);
+	if (!f) goto FAILURE;
+
+	if (makeflow_clean_file(d, q, f, 1, alloc)) {
+		debug(D_MAKEFLOW_RUN, "Unable to clean failed output");
+		goto FAILURE;
+	}
+	if (batch_fs_mkdir(q, f->filename, 0755, 0)) {
+		debug(D_MAKEFLOW_RUN, "Unable to create failed output directory: %s", strerror(errno));
+		goto FAILURE;
+	}
+
+	makeflow_log_file_state_change(d, f, DAG_FILE_STATE_COMPLETE);
+	fprintf(stderr, "rule %d failed, moving any outputs to %s\n",
+			n->nodeid, faildir);
+	rc = 0;
+FAILURE:
+	free(faildir);
+	return rc;
+}
+
+int makeflow_clean_failed_file(struct dag *d, struct dag_node *n, struct batch_queue *q, struct dag_file *f, int prep_failed, int silent) {
+	assert(d);
+	assert(n);
+	assert(q);
+	assert(f);
+
+	if (prep_failed) goto CLEANUP;
+
+	char *failout = string_format(
+			FAIL_DIR "/%s", n->nodeid, f->filename);
+	struct dag_file *o = dag_file_lookup_fail(d, q, failout);
+	if (o) {
+		if (batch_fs_rename(q, f->filename, o->filename) < 0) {
+			debug(D_MAKEFLOW_RUN, "Failed to rename %s -> %s: %s",
+					f->filename, o->filename, strerror(errno));
+		} else {
+			makeflow_log_file_state_change(d, f, DAG_FILE_STATE_DELETE);
+			debug(D_MAKEFLOW_RUN, "Renamed %s -> %s",
+					f->filename, o->filename);
+		}
+	} else {
+		fprintf(stderr, "Skipping rename %s -> %s", f->filename, failout);
+	}
+	free(failout);
+CLEANUP:
+	return makeflow_clean_file(d, q, f, silent, NULL);
+}
+
+int makeflow_hook_fail_dir_prep(struct dag *d, struct dag_node *n, struct batch_queue *q) {
+	assert(d);
+	assert(n);
+	assert(q);
+
+	int rc = MAKEFLOW_HOOK_FAILURE;
+	char *faildir = string_format(FAIL_DIR, n->nodeid);
+	struct dag_file *f = dag_file_lookup_fail(d, q, faildir);
+	if (!f) goto FAILURE;
+
+	if (makeflow_clean_file(d, q, f, 1, NULL)) {
+		debug(D_MAKEFLOW_HOOK, "Unable to clean failed output");
+		goto FAILURE;
+	}
+	if (mkdir(f->filename, 0755)) {
+		debug(D_MAKEFLOW_HOOK, "Unable to create failed output directory: %s", strerror(errno));
+		goto FAILURE;
+	}
+
+	makeflow_log_file_state_change(d, f, DAG_FILE_STATE_COMPLETE);
+	fprintf(stderr, "rule %d failed, moving any outputs to %s\n",
+			n->nodeid, faildir);
+	rc = MAKEFLOW_HOOK_SUCCESS;
+FAILURE:
+	free(faildir);
+	return rc;
+}
+
+static int node_success(struct dag_node *n, struct batch_task *task){
+	struct dag *d = n->d;
+	struct batch_queue *q = makeflow_get_remote_queue();
+
+	assert(d);
+	assert(n);
+	assert(q);
+
+	int rc = MAKEFLOW_HOOK_FAILURE;
+	char *faildir = string_format(FAIL_DIR, n->nodeid);
+	struct dag_file *f = dag_file_lookup_fail(d, q, faildir);
+	if (!f) goto OUT;
+
+	if (makeflow_clean_file(n->d, q, f, 1, NULL)) {
+		debug(D_MAKEFLOW_HOOK, "Unable to clean failed output");
+		goto OUT;
+	}
+
+	rc = MAKEFLOW_HOOK_SUCCESS;
+
+OUT:
+	free(faildir);
+	return rc;
+}
+
+static int node_fail(struct dag_node *n, struct batch_task *task){
+	struct batch_file *bf = NULL;
+	struct dag_file *df = NULL;
+	int prep_failed =  makeflow_hook_fail_dir_prep(n->d, n, task->queue); 
+	if (prep_failed) { 
+		fprintf(stderr, "rule %d failed, cannot move outputs\n", 
+					n->nodeid); 
+	}
+
+	/* Clean files created in node. Clean existing and expected and record deletion. */
+	list_first_item(task->output_files);
+	while((bf = list_next_item(task->output_files))) {
+		df = dag_file_lookup_or_create(n->d, bf->outer_name);
+		if(df->state == DAG_FILE_STATE_EXPECT) {
+			makeflow_clean_failed_file(n->d, n, makeflow_get_queue(n), df, prep_failed, 1);
+		} else {
+			makeflow_clean_failed_file(n->d, n, makeflow_get_queue(n), df, prep_failed, 0);
+		}
+	}
+
+	return MAKEFLOW_HOOK_SUCCESS;
+}
+
+struct makeflow_hook makeflow_hook_fail_dir = {
+	.module_name = "Fail Dir",
+
+	.node_success = node_success,
+	.node_fail = node_fail,
+
+};
+
+
